✨ feat: 추가 flow연결, RightPostureViewController생성, settingView임시삭제 (#62)
