+++ DdokBaro/ViewControllers/WelcomeViewController.swift
@@ -19,7 +19,6 @@ class WelcomeViewController: UIViewController {
     @IBOutlet weak var titleLabel: UILabel!
     @IBOutlet weak var turtleImage: UIImageView!
     @IBOutlet weak var descriptionLabel: UILabel!
-    @IBOutlet weak var startingButton: UIButton!
     
     @IBOutlet weak var chartButton: UIButton!
     
@@ -27,8 +26,9 @@ class WelcomeViewController: UIViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        let backBarButtonItem = UIBarButtonItem(title: "자세 측정", style: .plain, target: self, action: nil)
-        self.navigationItem.backBarButtonItem = backBarButtonItem
+//        navigationController?.isNavigationBarHidden = true
+//        let backBarButtonItem = UIBarButtonItem(title: "자세 측정", style: .plain, target: self, action: nil)
+//        self.navigationItem.backBarButtonItem = backBarButtonItem
         
         chartButton.circleButton = true
         chartButton.setImage(UIImage(systemName: "chart.bar.xaxis"), for: .normal)
@@ -59,13 +59,13 @@ class WelcomeViewController: UIViewController {
         self.view.sendSubviewToBack(welcomeTurtleView)
     }
     
-    @IBAction func goSettingButton(_ sender: UIButton) {
-        
-        let storyboard = UIStoryboard(name: "Main", bundle: nil)
-        if let settingViewController = storyboard.instantiateViewController(withIdentifier: "SettingViewController") as? SettingViewController {
-            navigationController?.pushViewController(settingViewController, animated: true)
-        }
-    }
+//    @IBAction func goSettingButton(_ sender: UIButton) {
+//        
+//        let storyboard = UIStoryboard(name: "Main", bundle: nil)
+//        if let settingViewController = storyboard.instantiateViewController(withIdentifier: "SettingViewController") as? SettingViewController {
+//            navigationController?.pushViewController(settingViewController, animated: true)
+//        }
+//    }
 
     func changeTextColor() {
         guard let text = self.titleLabel.text else {return}
